Fix cold-start intialization problem.
When start a new container through containerd client, the init process
and the app inside the container may not be initialized completely.
So the immediate request will failed (e.g, connection refused).

In this commit, add a retry mechanism: if failed then sleeping for 0.5
seconds, try agagin. If failed agagin, sleeping time will be
multiplied by 4.