Add many features.
This commit does not follow the principle of SoftwareEngineering :(

First I add the implementation of baseline (gc of idle containers).

Second I add the limitation to the cold start concurrency to prevent
OOM in baseline.

Third, I change the kill container logic (e.g., umount overlay and
notify containerd).

Fourth, I change the reap logic. Although goroutine is cheap in golang,
but too many goroutine call syscall wait() to reap child process is not
very efficient, because it will occupy an OS thread.

Finally, I choose to populate containers for our method. Because switch
is relatively cheap, so it does not need cleanup idle containers and we
can populate it with some hello-world dummy containers.