Garbage collect containers with goroutines and fix reap bugs.
In previous implementation, gc is too slow and block in one goroutine.
So even after waiting for gc interval, the container may still exist and
can be reuse, which leading to high but incorrect cache hit rate. So I
choose to kick it out of CtrPool as soon as possible and gc it in
background with seperate goroutines.

And in previous implementation, after reap one process (for switch), it
will not be removed and might be wait again. Fix this by remove it after
reaping successfully.