Add lru or priority queue base on activate time for containers.
In previous implementation, the kill instances or switch candidate are
selected based on random number + simple queue. For effenciency it is
not sufficient, so I choose to implement an lru container pool.

There are two lru list or priority qeueu: local one for each lambda and
a global one. The local one is used for reuse selection and the global
one is for switch and kill (to release more memory space).

One more thing is that add a delete interface for metric, since I add a
warmup phase in evaluation, so we need cleanup the metric generated at
warmup phase.